Added resource files, shader functions and texture classes/loaders.
Added multiple new functions to the Shader class that allows for complex operations to be done in just one line!
Added a texture class that loads in a texture from a .png file (More formats probably to come).
Added in code to CMake files that copies the resources to the build directory, which also means that they will now be included in the github repo! (You can actually build the app now :D)