Redid the render function in the graphics class and added Sprite classes aswell as a SpriteBatch.
The SpriteBatch contains a list of all the active sprites and orders them so i can then easily draw them all with one draw call in the Graphics class.
The new render function reads all the sprites from the SpriteBatch and generate DrawRanges that works as directives for when to change texture aswell as shader.

At the moment the program just draws all the sprites with the same texture and shader but in the short future the program will be able to draw using diffirent textures and shaders.
TODO:
    Expand on the DrawRange functionality